requirements.txt
# Core dependencies with flexible versioning for Python 3.10+ compatibility
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.1

# ML/AI dependencies - using more recent versions for Python 3.10+
numpy>=1.26.0
pandas>=2.1.0
sentence-transformers>=2.2.2

# Text processing
beautifulsoup4>=4.12.2
markdown>=3.5.1

# Vector database - using more recent version for Python 3.10+
faiss-cpu>=1.7.4
